# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: Pocket AI Gateway Management API
  version: 0.2.0
  description: Implemented Phase 2 management surface. Compatible inference APIs are specified separately when implemented.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
  /auth/setup/status:
    get:
      security: []
      operationId: setupStatus
      responses:
        "200":
          description: Safe setup state
          content:
            application/json:
              schema:
                type: object
                required: [setup_required, setup_recovery_available]
                properties:
                  setup_required: { type: boolean }
                  setup_recovery_available: { type: boolean }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /auth/setup/claim:
    post:
      security: []
      operationId: claimOwner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OwnerClaim" }
      responses:
        "201": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /auth/login:
    post:
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, format: password, maxLength: 1024, writeOnly: true }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /auth/activate:
    post:
      security: []
      operationId: activateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code, password]
              properties:
                code: { type: string, writeOnly: true }
                password: { type: string, format: password, minLength: 12, maxLength: 1024, writeOnly: true }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /auth/logout:
    post:
      operationId: logout
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      responses:
        "204": { description: Session revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /auth/session:
    get:
      operationId: currentSession
      responses:
        "200":
          description: Current user and session
          content:
            application/json:
              schema:
                type: object
                required: [user, session]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  session: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /auth/sessions:
    get:
      operationId: listSessions
      responses:
        "200":
          description: Owned sessions
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/Session" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /auth/sessions/{id}:
    delete:
      operationId: revokeSession
      parameters:
        - { $ref: "#/components/parameters/ID" }
        - { $ref: "#/components/parameters/CSRF" }
      responses:
        "204": { description: Session revoked }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /auth/password:
    post:
      operationId: changePassword
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password, writeOnly: true }
                new_password: { type: string, format: password, minLength: 12, maxLength: 1024, writeOnly: true }
      responses:
        "204": { description: Password changed; other sessions revoked }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
  /me:
    get:
      operationId: getProfile
      responses:
        "200": { $ref: "#/components/responses/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateProfile
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, display_name]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                display_name: { type: string, minLength: 1, maxLength: 100 }
                current_password: { type: string, format: password, writeOnly: true, description: Required when email changes }
      responses:
        "200": { $ref: "#/components/responses/User" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: pocket_ai_gateway_session }
  parameters:
    ID: { name: id, in: path, required: true, schema: { type: string } }
    CSRF: { name: X-CSRF-Token, in: header, required: true, schema: { type: string }, description: Must equal the readable strict SameSite CSRF cookie }
...
  responses:
    Error:
      description: Unexpected management API error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Invalid request
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Authentication failed or is required
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: The request is not permitted or could not be verified
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: Resource state conflict
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    PayloadTooLarge:
      description: Request body exceeds 64 KiB
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    UnsupportedMediaType:
      description: Content-Type must be application/json
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    TooManyRequests:
      description: Too many authentication attempts
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ServiceUnavailable:
      description: Required storage is unavailable
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    User:
      description: User resource
      content:
        application/json:
          schema:
            type: object
            required: [user]
            properties:
              user: { $ref: "#/components/schemas/User" }
...
  schemas:
    OwnerClaim:
      type: object
      additionalProperties: false
      required: [setup_code, email, display_name, password]
      properties:
        setup_code: { type: string, writeOnly: true }
        email: { type: string, format: email, maxLength: 254 }
        display_name: { type: string, minLength: 1, maxLength: 100 }
        password: { type: string, format: password, minLength: 12, maxLength: 1024, writeOnly: true }
    User:
      type: object
      required: [id, email, display_name, role, status, grants]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [owner, admin, member] }
        status: { type: string, enum: [active, suspended, pending_activation, archived] }
        grants: { $ref: "#/components/schemas/Grants" }
...
    Grants:
      type: object
      additionalProperties: false
      required: [unrestricted, scopes, model_patterns, connection_ids]
      properties:
        unrestricted: { type: boolean, description: Read-only true for the owner; user writes are forced false }
        scopes: { type: array, maxItems: 100, uniqueItems: true, items: { type: string, enum: [chat:generate, responses:generate, embeddings:generate, models:read, tokens:count] } }
        model_patterns: { type: array, maxItems: 100, uniqueItems: true, items: { type: string, minLength: 1, maxLength: 200 } }
        connection_ids: { type: array, maxItems: 100, uniqueItems: true, items: { type: string, minLength: 1, maxLength: 200 } }
    Session:
      type: object
      required: [id, current, created_at, last_seen_at, expires_at, authenticated_at, user_agent]
      properties:
        id: { type: string }
        current: { type: boolean }
        created_at: { type: string, format: date-time }
        last_seen_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        authenticated_at: { type: string, format: date-time }
        user_agent: { type: string }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }