# internal/storage/sqlc/systemdb/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package systemdb

import (
	"database/sql"
)

type ActivationToken struct {
	Verifier  []byte `json:"verifier"`
	UserID    string `json:"user_id"`
	Purpose   string `json:"purpose"`
	ExpiresAt int64  `json:"expires_at"`
	CreatedAt int64  `json:"created_at"`
}

type ApiKey struct {
	ID                string        `json:"id"`
	OwnerUserID       string        `json:"owner_user_id"`
	Label             string        `json:"label"`
	State             string        `json:"state"`
	ScopesJson        string        `json:"scopes_json"`
	ModelPatternsJson string        `json:"model_patterns_json"`
	ConnectionIdsJson string        `json:"connection_ids_json"`
	ExpiresAt         sql.NullInt64 `json:"expires_at"`
	Revision          int64         `json:"revision"`
	CreatedAt         int64         `json:"created_at"`
	UpdatedAt         int64         `json:"updated_at"`
}

type ApiKeySecret struct {
	Selector  string        `json:"selector"`
	Verifier  []byte        `json:"verifier"`
	KeyID     string        `json:"key_id"`
	CreatedAt int64         `json:"created_at"`
	ExpiresAt sql.NullInt64 `json:"expires_at"`
	RevokedAt sql.NullInt64 `json:"revoked_at"`
}

type AuditEvent struct {
	ID           string         `json:"id"`
	ActorUserID  sql.NullString `json:"actor_user_id"`
	Action       string         `json:"action"`
	ResourceType string         `json:"resource_type"`
	ResourceID   string         `json:"resource_id"`
	DetailJson   string         `json:"detail_json"`
	CreatedAt    int64          `json:"created_at"`
}

type GatewayMetadatum struct {
	Key       string `json:"key"`
	Value     string `json:"value"`
	UpdatedAt string `json:"updated_at"`
}

type LoginThrottle struct {
	Subject         []byte `json:"subject"`
	WindowStartedAt int64  `json:"window_started_at"`
	Failures        int64  `json:"failures"`
	BlockedUntil    int64  `json:"blocked_until"`
}

type SetupToken struct {
	Singleton     int64          `json:"singleton"`
	Verifier      []byte         `json:"verifier"`
	ExpiresAt     int64          `json:"expires_at"`
	ClaimedUserID sql.NullString `json:"claimed_user_id"`
}

type User struct {
	ID                    string `json:"id"`
	Email                 string `json:"email"`
	DisplayName           string `json:"display_name"`
	PasswordHash          string `json:"password_hash"`
	Role                  string `json:"role"`
	Status                string `json:"status"`
	CreatedAt             int64  `json:"created_at"`
	UpdatedAt             int64  `json:"updated_at"`
	AuthRevision          int64  `json:"auth_revision"`
	Revision              int64  `json:"revision"`
	InferenceUnrestricted int64  `json:"inference_unrestricted"`
	ScopesJson            string `json:"scopes_json"`
	ModelPatternsJson     string `json:"model_patterns_json"`
	ConnectionIdsJson     string `json:"connection_ids_json"`
}

type UserSession struct {
	ID              string `json:"id"`
	Verifier        []byte `json:"verifier"`
	UserID          string `json:"user_id"`
	ExpiresAt       int64  `json:"expires_at"`
	CreatedAt       int64  `json:"created_at"`
	LastSeenAt      int64  `json:"last_seen_at"`
	AuthenticatedAt int64  `json:"authenticated_at"`
	AuthRevision    int64  `json:"auth_revision"`
	UserAgent       string `json:"user_agent"`
}